+++ Assets/Scripts/InteractableObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4779f1fa9414aab49ac7ac120e99e897
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
