+++ Assets/Scripts/InteractableObject/IInteractable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: feef6437dcbf0974d854c5ab89509a75
+guid: 7b82bc6990b029e4783cac540f612280
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
