+++ Assets/Scripts/InteractableObject/InteractorBehavior.cs
@@ -1,28 +1,19 @@
+// Ignore Spelling: Interactor
+
 using System.Collections;
 using System.Collections.Generic;
-using System.Reflection.Emit;
-using UnityEditor;
 using UnityEngine;
 
-public interface IInteractable
-{
-
-    public void Focus();
-
-    public void Unfocus();
-
-    public void Select();
-
-    public void Deselect();
-
-    public void Interact();
-}
-
-public class InteractionSystem : MonoBehaviour
+public class InteractorBehavior : MonoBehaviour
 {
     private IInteractable selected;
     private Transform selectedTransform;
     private float selectedDistance = float.PositiveInfinity;
+
+    [Tooltip("Physics Layer to trigger focusing on an object")]
+    public PhysicsLayer_SO focusLayer;
+    [Tooltip("Physics Layer for selecting and interacting with the closest object")]
+    public PhysicsLayer_SO selectLayer;
 
     /// <summary>
     /// If the provided Collider object is closer than the currently active one, "select"
@@ -30,7 +21,7 @@ public class InteractionSystem : MonoBehaviour
     /// </summary>
     /// <param name="other">Collider to use for distance check</param>
     private void SelectIfCloser(Collider other)
-    { 
+    {
         float otherDistance = Vector3.Distance(transform.position, other.transform.position);
         if (otherDistance < selectedDistance)
         {
@@ -45,13 +36,12 @@ public class InteractionSystem : MonoBehaviour
 
     /// <summary>Update Focus/Selection when entering Collider.</summary>
     private void OnTriggerEnter(Collider other)
-    {
-        if (other.gameObject.layer == LayerMask.NameToLayer("Interaction Focus"))
-        {
-            other.gameObject.GetComponentInParent<IInteractable>()?.Focus();
+    { 
+        if (other.gameObject.layer == focusLayer.LayerIndex) { 
+            other.gameObject.GetComponentInParent<IInteractable>()?.Focus(); 
         }
 
-        else if (other.gameObject.layer == LayerMask.NameToLayer("Interaction Select"))
+        else if (other.gameObject.layer == selectLayer.LayerIndex)
         {
             var iinteractable = other.gameObject.GetComponentInParent<IInteractable>();
             if (iinteractable != null) SelectIfCloser(other);
@@ -61,7 +51,7 @@ public class InteractionSystem : MonoBehaviour
     /// <summary>Update Focus/Selection when staying within Collider.</summary>
     private void OnTriggerStay(Collider other)
     {
-        if (other.gameObject.layer == LayerMask.NameToLayer("Interaction Select"))
+        if (other.gameObject.layer == selectLayer.LayerIndex)
         {
             var iinteractable = other.gameObject.GetComponentInParent<IInteractable>();
             if (iinteractable != null) SelectIfCloser(other);
@@ -71,12 +61,12 @@ public class InteractionSystem : MonoBehaviour
     /// <summary>Update Focus/Selection when exiting Collider.</summary
     private void OnTriggerExit(Collider other)
     {
-        if (other.gameObject.layer == LayerMask.NameToLayer("Interaction Focus"))
+        if (other.gameObject.layer == focusLayer.LayerIndex)
         {
             other.gameObject.GetComponentInParent<IInteractable>()?.Unfocus();
         }
 
-        else if (other.gameObject.layer == LayerMask.NameToLayer("Interaction Select"))
+        else if (other.gameObject.layer == selectLayer.LayerIndex)
         {
             var iinteractable = other.gameObject.GetComponentInParent<IInteractable>();
             if (iinteractable == selected)
@@ -94,9 +84,12 @@ public class InteractionSystem : MonoBehaviour
     private void FixedUpdate()
     {
         if (selectedTransform != null)
-        { 
+        {
             selectedDistance = Vector3.Distance(transform.position, selectedTransform.position);
         }
     }
 
+    /// <summary>Called <c>.Interact()</c> on the currently selected object.</summary>
+    public void OnInteractPressed() { selected?.Interact(); }
 }
+
